tests: Common players
misc fixes of small errors that were found

+++ src/main/kotlin/cl/uchile/dcc/finalreality/model/character/Enemy.kt
@@ -45,5 +45,5 @@ class Enemy(
         return "$superString, weight=$weight)"
     }
 
-    override fun delay(): Int = weight / 10
+    override fun delay(): Long = (weight / 10).toLong()
 }
